requirements.txt
cachetools==4.1.0
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
cryptography==2.9.2
Django==2.2
django-extra-fields==2.0.5
django-gmailapi-backend==0.1.0
django-rest-knox==4.1.0
djangorestframework==3.11.0
google-api-python-client==1.7.11
google-auth==1.11.2
google-auth-httplib2==0.0.3
httpie==2.1.0
httplib2==0.17.3
idna==2.9
numpy==1.18.4
pandas==1.0.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pycurl==7.43.0.5
Pygments==2.6.1
PyJWT==1.7.1
python-dateutil==2.8.1
python-dotenv==0.13.0
pytz==2020.1
requests==2.23.0
rest-pandas==1.1.0
rope==0.17.0
rsa==4.0
six==1.14.0
sqlparse==0.3.1
uritemplate==3.0.1
urllib3==1.25.9
Pillow==7.1.2
openpyxl==3.0.3
